+++ src/index.js
@@ -1,5 +1,2 @@
-import * as cursor from './cursor';
-
 export default from './Editor';
 export * from './utils';
-export { cursor };