+++ src/schemas/equation.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { renderHtmlChildren } from '../utils/schemaUtils';
+import { renderHtmlChildren } from '../utils/render';
 
 export default {
 	equation: {